box weight data type changes

+++ src/main/java/com/sericulture/masterdata/model/api/marketMaster/EditMarketMasterRequest.java
@@ -30,7 +30,7 @@ public class EditMarketMasterRequest extends RequestBody {
     String marketMasterAddress;
 
     @Schema(name = "boxWeight", example = "10")
-    Integer boxWeight;
+    BigDecimal boxWeight;
 
     @Schema(name = "lotWeight", example = "5")
     Integer lotWeight;
